added types to account settings

+++ src/User/Account/AccountSettings.tsx
@@ -1,18 +1,26 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, ChangeEvent } from 'react';
 import { Link } from 'react-router-dom';
 import fireDB from "../../firebaseConfig";
 import { deleteDoc, doc, getDoc, setDoc } from "firebase/firestore";
-import { EmailAuthProvider, getAuth, reauthenticateWithCredential, updateEmail, deleteUser } from "firebase/auth";
+import { EmailAuthProvider, getAuth, reauthenticateWithCredential, updateEmail, deleteUser, User, EmailAuthCredential } from "firebase/auth";
 import { logout } from '../../App';
+
+interface Values {
+ name: any;
+ email: any;
+ password: string;
+ delete: string;
+ reason: string;
+};
 
 export default function AccountSettings() {
 
- const user = JSON.parse(localStorage.getItem("currentUser"));
+ const user = JSON.parse(localStorage.getItem("currentUser") || "{}");
 
- const initialValues = { name: user.userInfo.name, email: user.email, password: '', format: user.userInfo.format, delete: '', reason: '' };
+ const initialValues = { name: user.userInfo.name, email: user.email, password: '', delete: '', reason: '' };
 
  
- const [values, setValues] = useState(initialValues);
+ const [values, setValues] = useState<Values>(initialValues);
  
  useEffect(() => {
   console.clear();
@@ -21,14 +29,22 @@ export default function AccountSettings() {
 
  const auth = getAuth();
 
- const handleChange = e => {
+ const handleChange = (e: any): void => {
   setValues({
    ...values,
    [e.target.name]: e.target.value
   });
  }
 
- const saveChanges = async () => {
+ const [format, setFormat] = useState<any>(user.userInfo.format);
+
+ const onRadioChange = (e: ChangeEvent<HTMLInputElement>): void => {
+  setFormat(e.target.value);
+ }
+
+ const isRadioSelected = (value: string): boolean => format === value;
+
+ const saveChanges = async (): Promise<any> => {
   if (values.password === '') {
    alert("You have to enter your password in the password field to save your changes");
    return;
@@ -39,26 +55,30 @@ export default function AccountSettings() {
    const userInfo = {
     ...userDoc.data(),
     name: values.name,
-    format: values.format
+    format
    };
    await setDoc(docRef, userInfo);
    localStorage.setItem("currentUser", JSON.stringify({...user, userInfo}));
-   const currentUser = auth.currentUser;
-   const credential = EmailAuthProvider.credential(
-    auth.currentUser.email,
-    values.password
-   );
-   await reauthenticateWithCredential(currentUser, credential);
-   await updateEmail(currentUser, `${values.email}`);
-   localStorage.setItem("currentUser", JSON.stringify({...JSON.parse(localStorage.getItem("currentUser")), email: `${values.email}`}));
+   const currentUser: User | null = auth.currentUser;
+   if (currentUser) {
+    const credential: EmailAuthCredential = EmailAuthProvider.credential(
+     currentUser.email!,
+     values.password
+    );
+    await reauthenticateWithCredential(currentUser, credential);
+    await updateEmail(currentUser, `${values.email}`);
+   } else {
+    return;
+   }
+   localStorage.setItem("currentUser", JSON.stringify({...JSON.parse(localStorage.getItem("currentUser") || "{}"), email: `${values.email}`}));
    alert("Changes saved");
-   window.location = "/account";
+   window.location.href = "/account";
   } catch (err) {
    alert(`Change saving error: ${err}`);
   }
  }
 
- const deleteAccount = async () => {
+ const deleteAccount = async (): Promise<any> => {
   if (values.delete !== "DELETE") {
    alert("Type in DELETE to delete your account");
    return;
@@ -78,13 +98,17 @@ export default function AccountSettings() {
    await setDoc(deletedUserRef, deletedUser);
    const userDocRef = doc(fireDB, "users", `${user.uid}`);
    await deleteDoc(userDocRef);
-   const currentUser = auth.currentUser;
-   const credential = EmailAuthProvider.credential(
-    auth.currentUser.email,
-    values.password
-   );
-   await reauthenticateWithCredential(currentUser, credential);
-   await deleteUser(currentUser);
+   const currentUser: User | null = auth.currentUser;
+   if (currentUser) {
+    const credential = EmailAuthProvider.credential(
+     currentUser.email!,
+     values.password
+    );
+    await reauthenticateWithCredential(currentUser, credential);
+    await deleteUser(currentUser);
+   } else {
+    return;
+   }
    alert("Account deleted");
    logout();
   } catch (err) {
@@ -105,9 +129,23 @@ export default function AccountSettings() {
    <Link to="/account/change_password"><button>Change password</button></Link>
    <div style={{display: "flex", flexDirection: "column"}}>
     <h4>Time format:</h4>
-    <div value={values.format} onChange={handleChange}>
-     <input type="radio" name="format" checked={values.format === "ampm"} value="ampm"/><label>AM/PM Format</label>
-     <input type="radio" name="format" checked={values.format === "MT"} value="MT"/><label>24 Hr. Format</label>
+    <div>
+     <input 
+      type="radio" 
+      name="format" 
+      value="ampm"
+      checked={isRadioSelected("ampm")} 
+      onChange={onRadioChange}
+     />
+     <label>AM/PM Format</label>
+     <input 
+      type="radio" 
+      name="format" 
+      value="MT"
+      checked={isRadioSelected("MT")} 
+      onChange={onRadioChange}
+     />
+     <label>24 Hr. Format</label>
     </div>
     <h4>Location privacy:</h4>
     <input type="radio" name="location"/><label>Show my location to anyone</label>
@@ -124,7 +162,7 @@ export default function AccountSettings() {
     <label>Type DELETE to confirm:</label>
     <input type="text" name="delete" value={values.delete} onChange={handleChange}/>
     <label>Why are you deleting your account? (Optional)</label>
-    <textarea rows="10" cols="10" name="reason" value={values.reason} onChange={handleChange}/>
+    <textarea rows={10} cols={10} name="reason" value={values.reason} onChange={handleChange}/>
    </div>
    <button onClick={deleteAccount}>Delete my account</button>
   </div>
